+++ source/core/php/index.js
@@ -20,6 +20,7 @@ class PHP extends Base {
       'database/mssql',
       'database/sqlsrv',
       'database/oracle',
+      'database/oracle_oci8',
       'database/informix'
     ].map((_) => {
       this.parseTemplate(`./php/template/${_}`);